Created Schema for SwiftData
