+++ DisHabitApp/View/Components/QuestBoardView.swift
@@ -24,11 +24,11 @@ struct QuestBoardView: View {
         }
     }
     
-    @Query var standbyQuests: [Quest]
-    @Query var dailyQuestBoards: [DailyQuestBoard]
-    @Query var objectives: [Objective] //temp
-    @Query var tasks: [StandbyTask] //temp!!
-    @Query var qs: [QuestSlot] //temp!!
+    @Query var standbyQuests: [SchemaV1.Quest]
+    @Query var dailyQuestBoards: [SchemaV1.DailyQuestBoard]
+    @Query var objectives: [SchemaV1.Objective] //temp
+    @Query var tasks: [SchemaV1.StandbyTask] //temp!!
+    @Query var qs: [SchemaV1.QuestSlot] //temp!!
     
     
     
@@ -63,8 +63,8 @@ struct QuestBoardView: View {
         // activeQuestsの中から、selectedDateの曜日を持つQuestからQuestSlotを作成
         let date = self.selectedDate
         
-        let newBoard = DailyQuestBoard(date: date, questSlots: [])
-        let questSlots = standbyQuests.filter { $0.activatedDayOfWeeks[date.weekday()] == true }.map { QuestSlot(board: newBoard, quest: $0) }
+        let newBoard = SchemaV1.DailyQuestBoard(date: date, questSlots: [])
+        let questSlots = standbyQuests.filter { $0.activatedDayOfWeeks[date.weekday()] == true }.map { SchemaV1.QuestSlot(board: newBoard, quest: $0) }
         newBoard.questSlots = questSlots
 
         
@@ -122,6 +122,7 @@ struct QuestBoardView: View {
                         for board in dailyQuestBoards {
                             print(board.date)
                         }
+                        print(qs.count)
                         _Concurrency.Task {
 //                            await vm.debug_ResetAcceptedQuests()
                         }
@@ -130,43 +131,43 @@ struct QuestBoardView: View {
                     }
                     Button {
                         do {
-                            try modelContext.delete(model: DailyQuestBoard.self)
-                            try modelContext.delete(model: DailyQuestBoard.self)
+                            try modelContext.delete(model: SchemaV1.DailyQuestBoard.self)
+                            try modelContext.delete(model: SchemaV1.DailyQuestBoard.self)
 //                            let a = tasks
 //                            let b = objectives
 //                            let c = standbyQuests
                             
-//                            let objective1 = Objective(id: UUID(), text: "健康的な生活を送る")
-//                            let objective2 = Objective(id: UUID(), text: "勉強習慣を身につける")
-//                            let objective3 = Objective(id: UUID(), text: "運動を習慣化する")
-//                    
-//                            let task1 = StandbyTask(id: UUID(), text: "朝7時に起床する", objective: objective1)
-//                            let task2 = StandbyTask(id: UUID(), text: "朝食を食べる", objective: objective1)
-//                            let task3 = StandbyTask(id: UUID(), text: "1時間勉強する", objective: objective2)
-//                            let task4 = StandbyTask(id: UUID(), text: "30分ジョギングする", objective: objective3)
-//                            let task5 = StandbyTask(id: UUID(), text: "ストレッチをする", objective: objective3)
-//                    
-//                            let reward1 = Reward(id: UUID(), text: "好きなお菓子を1つ買う")
-//                            let reward2 = Reward(id: UUID(), text: "映画を見る")
-//                    
-//                            let quest1 = Quest(id: UUID(), activatedDayOfWeeks: [1: true, 2: true, 3: true, 4: true, 5: true, 6: true, 7: true], reward: reward1, tasks: [task1, task2, task5])
-//                            let quest2 = Quest(id: UUID(), activatedDayOfWeeks: [1: true, 2: true, 3: true, 4: true, 5: true, 6: true, 7: true], reward: reward2, tasks: [task4, task5])
-//                            
-//                        
-//                            modelContext.insert(objective1)
-//                            modelContext.insert(objective2)
-//                            modelContext.insert(objective3)
-//                            modelContext.insert(task1)
-//                            modelContext.insert(task2)
-//                            modelContext.insert(task3)
-//                            modelContext.insert(task4)
-//                            modelContext.insert(task5)
-//                            modelContext.insert(reward1)
-//                            modelContext.insert(reward2)
-//                            modelContext.insert(quest1)
-//                            modelContext.insert(quest2)
-//                            
-//                            try modelContext.save()
+                            let objective1 = SchemaV1.Objective(id: UUID(), text: "健康的な生活を送る")
+                            let objective2 = SchemaV1.Objective(id: UUID(), text: "勉強習慣を身につける")
+                            let objective3 = SchemaV1.Objective(id: UUID(), text: "運動を習慣化する")
+                    
+                            let task1 = SchemaV1.StandbyTask(id: UUID(), text: "朝7時に起床する", objective: objective1)
+                            let task2 = SchemaV1.StandbyTask(id: UUID(), text: "朝食を食べる", objective: objective1)
+                            let task3 = SchemaV1.StandbyTask(id: UUID(), text: "1時間勉強する", objective: objective2)
+                            let task4 = SchemaV1.StandbyTask(id: UUID(), text: "30分ジョギングする", objective: objective3)
+                            let task5 = SchemaV1.StandbyTask(id: UUID(), text: "ストレッチをする", objective: objective3)
+                    
+                            let reward1 = SchemaV1.Reward(id: UUID(), text: "好きなお菓子を1つ買う")
+                            let reward2 = SchemaV1.Reward(id: UUID(), text: "映画を見る")
+                    
+                            let quest1 = SchemaV1.Quest(id: UUID(), activatedDayOfWeeks: [1: true, 2: true, 3: true, 4: true, 5: true, 6: true, 7: true], reward: reward1, tasks: [task1, task2, task5])
+                            let quest2 = SchemaV1.Quest(id: UUID(), activatedDayOfWeeks: [1: true, 2: true, 3: true, 4: true, 5: true, 6: true, 7: true], reward: reward2, tasks: [task4, task5])
+                            
+                        
+                            modelContext.insert(objective1)
+                            modelContext.insert(objective2)
+                            modelContext.insert(objective3)
+                            modelContext.insert(task1)
+                            modelContext.insert(task2)
+                            modelContext.insert(task3)
+                            modelContext.insert(task4)
+                            modelContext.insert(task5)
+                            modelContext.insert(reward1)
+                            modelContext.insert(reward2)
+                            modelContext.insert(quest1)
+                            modelContext.insert(quest2)
+                            
+                            try modelContext.save()
                             
                             print("Deleted board & qs table")
                         } catch {
@@ -176,10 +177,10 @@ struct QuestBoardView: View {
                         Text("Board/QSテーブルをリセット")
                     }
 #endif
-//                    ForEach(displayedQuestSlots.indices, id: \.self) { index in
-//                        QuestSlotContainer(manager: displayedQuestSlots[index])
-//
-//                    }
+                    ForEach(displayedQuestSlots.indices, id: \.self) { index in
+                        QuestSlotContainer(manager: displayedQuestSlots[index])
+
+                    }
 
                 }
             }
