+++ DisHabitApp/View/Components/TicketCard.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 struct TicketCard: View {
 //    @ObservedObject var vm: QuestBoardViewModel
-    var acceptedQuest: AcceptedQuest
+    var acceptedQuest: SchemaV1.AcceptedQuest
     
     var body: some View {
         if acceptedQuest.reward.isRedeemed {
